Move TreePath functions into their own file
